move all startup and shutdown into lifespan app
